+++ AIDevGallery/Samples/Open Source Models/Stable Diffusion/GenerateImage.xaml.cs
@@ -146,7 +146,7 @@ namespace AIDevGallery.Samples.OpenSourceModels.StableDiffusionImageGeneration
                         {
                             this.DispatcherQueue.TryEnqueue(() =>
                             {
-                                BitmapImage bitmapImage = BitmapFunctions.ConvertBitmapToBitmapImageAsync(image);
+                                BitmapImage bitmapImage = BitmapFunctions.ConvertBitmapToBitmapImage(image);
                                 DefaultImage.Source = bitmapImage;
                                 NarratorHelper.AnnounceImageChanged(DefaultImage, "Image changed: new image generated."); // <exclude-line>
                                 DefaultImage.Visibility = Visibility.Visible;
